adjmig

+++ database/migrations/2022_06_21_180328_comuna.php
@@ -16,7 +16,7 @@ class Comuna extends Migration
         Schema::create('comuna', function (Blueprint $table) {
             $table->id();
             $table->string('nombre')->nullable();
-            $table->integer('ciudad_id')->unsigned()->index()->nullable();
+            $table->bigInteger('ciudad_id')->unsigned()->index()->nullable();
             $table->foreign('ciudad_id')->references('id')->on('ciudad')->onDelete('cascade');
 
         });
